requirements.txt
Django==1.11.7
django-registration-redux==1.4
Pillow==4.0.0